It works!!!!!

+++ src/systems/systems/src/simulation/mod.rs
@@ -21,8 +21,8 @@ pub trait SimulatorReaderWriter {
 
     fn update_brake_input_left(&mut self, left_raw_val: u32);
     fn update_brake_input_right(&mut self, right_raw_val: u32);
-    fn get_brake_output_left(&mut self) -> f64;
-    fn get_brake_output_right(&mut self) -> f64;
+    fn get_brake_output_left(&mut self) -> u32;
+    fn get_brake_output_right(&mut self) -> u32;
 }
 
 /// An [`Aircraft`] that can be simulated by the [`Simulation`].
@@ -252,11 +252,11 @@ impl<'a, T: Aircraft, U: SimulatorReaderWriter> Simulation<'a, T, U> {
             .update_brake_input_right(raw_right_val);
     }
 
-    pub fn get_brake_output_left(&mut self) -> f64 {
+    pub fn get_brake_output_left(&mut self) -> u32 {
         self.simulator_read_writer.get_brake_output_left()
     }
 
-    pub fn get_brake_output_right(&mut self) -> f64 {
+    pub fn get_brake_output_right(&mut self) -> u32 {
         self.simulator_read_writer.get_brake_output_right()
     }
 }
